--- app/(authenticated)/layout.tsx
@@ -1,41 +1,0 @@
-"use client";
-
-import { useAuthenticator } from "@aws-amplify/ui-react";
-import { useRouter } from "next/navigation";
-import { useEffect } from "react";
-
-export default function AuthenticatedLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const { authStatus, user } = useAuthenticator();
-  const router = useRouter();
-
-  useEffect(() => {
-    // If not authenticated, redirect to login
-    if (authStatus === "unauthenticated") {
-      router.push("/login");
-    }
-  }, [authStatus, router]);
-
-  // Show loading while checking authentication
-  if (authStatus === "configuring") {
-    return (
-      <div className="min-h-screen bg-gray-100 flex items-center justify-center">
-        <div className="text-center">
-          <div className="animate-spin rounded-full h-32 w-32 border-b-2 border-blue-500 mx-auto"></div>
-          <p className="mt-4 text-gray-600">Loading...</p>
-        </div>
-      </div>
-    );
-  }
-
-  // If not authenticated, don't render children
-  if (authStatus === "unauthenticated") {
-    return null;
-  }
-
-  // If authenticated, render children
-  return <>{children}</>;
-}
