+++ app/layout.tsx
@@ -2,21 +2,14 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
 
-import AuthenticatorWrapper from "./AuthenticatorWrapper";
 import QueryProvider from "@/components/providers/QueryProvider";
-import "@aws-amplify/ui-react/styles.css";
-
-import { Amplify } from "aws-amplify";
-import outputs from "../amplify_outputs.json";
 
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "[TODO]",
-  description: "[TODO]",
+  title: "AI-Powered Document Analysis",
+  description: "Upload PDFs and ask questions to get AI-powered answers",
 };
-
-Amplify.configure(outputs);
 
 export default function RootLayout({
   children,
@@ -26,9 +19,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>
-        <QueryProvider>
-          <AuthenticatorWrapper>{children}</AuthenticatorWrapper>
-        </QueryProvider>
+        <QueryProvider>{children}</QueryProvider>
       </body>
     </html>
   );
