+++ middleware.ts
@@ -2,30 +2,7 @@ import { NextResponse } from 'next/server'
 import type { NextRequest } from 'next/server'
 
 export function middleware(request: NextRequest) {
-  // Get the pathname of the request
-  const path = request.nextUrl.pathname
-
-  // Define public routes that don't require authentication
-  const publicRoutes = ['/', '/login']
-  
-  // Check if the current path is a public route
-  const isPublicRoute = publicRoutes.includes(path)
-  
-  // Get the authentication token from cookies
-  const authToken = request.cookies.get('amplify-authenticator-authToken')?.value
-  
-  // If it's not a public route and user is not authenticated, redirect to login
-  if (!isPublicRoute && !authToken) {
-    const loginUrl = new URL('/login', request.url)
-    return NextResponse.redirect(loginUrl)
-  }
-  
-  // If user is authenticated and trying to access login page, redirect to dashboard
-  if (path === '/login' && authToken) {
-    const dashboardUrl = new URL('/dashboard', request.url)
-    return NextResponse.redirect(dashboardUrl)
-  }
-  
+  // Simple pass-through middleware - no authentication required
   return NextResponse.next()
 }
